+++ src/cloudflare.ts
@@ -16,6 +16,7 @@ export function createD1(config: RuntimeConfig): RuntimeConfig {
       'd1',
       'create',
       config.d1DatabaseName,
+      '--update-config=false',
     ],
     config
   );
@@ -41,6 +42,7 @@ export function createR2(config: RuntimeConfig): void {
       'bucket',
       'create',
       config.r2BucketName,
+      '--update-config=false',
     ],
     config
   );
@@ -51,7 +53,15 @@ export function createKV(config: RuntimeConfig): RuntimeConfig {
 
   const result = runCommand(
     'pnpm',
-    ['exec', 'wrangler', 'kv', 'namespace', 'create', config.kvNamespaceName],
+    [
+      'exec',
+      'wrangler',
+      'kv',
+      'namespace',
+      'create',
+      config.kvNamespaceName,
+      '--update-config=false',
+    ],
     config
   );
   const outputText = `${result.stdout}\n${result.stderr}`;